requirements.txt
# Activating venv
# python -m venv SPDMenv
# SPDMenv\Scripts\activate

# pip install -r requirements.txt

# Starting Flask app
# python app.py

# comandi utili

# forzare la creazione dei DB:
# nel terminale
# python
# from app import app, db
# with app.app_context(): db.create_all()
# exit

# pulire la cash e terminare i processi in atto
# Get-Process python* | Select-Object Id,ProcessName
# Get-Process python* | Stop-Process -Force

flask==3.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
flask_login==0.6.3
python-dotenv==1.0.1
flask_bcrypt==1.0.1
pillow==11.1.0



